Spread REST endpoints and methods types across EE code
