+++ apps/meteor/ee/app/api-enterprise/server/canned-responses.ts
@@ -1,9 +1,45 @@
 import { Meteor } from 'meteor/meteor';
 import { isPOSTCannedResponsesProps, isDELETECannedResponsesProps, isCannedResponsesProps } from '@rocket.chat/rest-typings';
+import type { ILivechatDepartment, IOmnichannelCannedResponse, IUser } from '@rocket.chat/core-typings';
+import type { PaginatedResult, PaginatedRequest } from '@rocket.chat/rest-typings';
 
 import { API } from '../../../../app/api/server';
 import { findAllCannedResponses, findAllCannedResponsesFilter, findOneCannedResponse } from './lib/canned-responses';
 import { getPaginationItems } from '../../../../app/api/server/helpers/getPaginationItems';
+
+declare module '@rocket.chat/rest-typings' {
+	// eslint-disable-next-line @typescript-eslint/naming-convention
+	interface Endpoints {
+		'/v1/canned-responses': {
+			GET: (
+				params: PaginatedRequest<{
+					shortcut?: string;
+					text?: string;
+					scope?: string;
+					createdBy?: IUser['username'];
+					tags?: any;
+					departmentId?: ILivechatDepartment['_id'];
+				}>,
+			) => PaginatedResult<{
+				cannedResponses: IOmnichannelCannedResponse[];
+			}>;
+			POST: (params: {
+				_id?: IOmnichannelCannedResponse['_id'];
+				shortcut: string;
+				text: string;
+				scope: string;
+				tags?: any;
+				departmentId?: ILivechatDepartment['_id'];
+			}) => void;
+			DELETE: (params: { _id: IOmnichannelCannedResponse['_id'] }) => void;
+		};
+		'/v1/canned-responses/:_id': {
+			GET: () => {
+				cannedResponse: IOmnichannelCannedResponse;
+			};
+		};
+	}
+}
 
 API.v1.addRoute(
 	'canned-responses.get',
