+++ apps/meteor/ee/app/livechat-enterprise/server/api/business-hours.ts
@@ -1,6 +1,22 @@
+import type { ILivechatBusinessHour } from '@rocket.chat/core-typings';
+
 import { API } from '../../../../../app/api/server';
 import { getPaginationItems } from '../../../../../app/api/server/helpers/getPaginationItems';
 import { findBusinessHours } from '../business-hour/lib/business-hour';
+
+declare module '@rocket.chat/rest-typings' {
+	// eslint-disable-next-line @typescript-eslint/naming-convention
+	interface Endpoints {
+		'/v1/livechat/business-hours': {
+			GET: (params: { name?: string; offset: number; count: number; sort: Record<string, unknown> }) => {
+				businessHours: ILivechatBusinessHour[];
+				count: number;
+				offset: number;
+				total: number;
+			};
+		};
+	}
+}
 
 API.v1.addRoute(
 	'livechat/business-hours',
