+++ apps/meteor/ee/client/omnichannel/cannedResponses/CannedResponseEditWithData.tsx
@@ -6,7 +6,6 @@ import React from 'react';
 import { FormSkeleton } from '../../../../client/components/Skeleton';
 import { AsyncStatePhase } from '../../../../client/hooks/useAsyncState';
 import { useEndpointData } from '../../../../client/hooks/useEndpointData';
-import '../../../definition/rest';
 import CannedResponseEdit from './CannedResponseEdit';
 import CannedResponseEditWithDepartmentData from './CannedResponseEditWithDepartmentData';
