--- apps/meteor/ee/definition/.eslintrc.json
@@ -1,3 +1,0 @@
-{
-	"extends": "../../client/.eslintrc.json"
-}
