--- apps/meteor/ee/definition/index.ts
@@ -1,2 +1,0 @@
-import './methods';
-import './rest';
