--- apps/meteor/ee/definition/methods/index.ts
@@ -1,1 +1,0 @@
-import './license';
