--- apps/meteor/ee/definition/rest/index.ts
@@ -1,7 +1,0 @@
-// TODO: chapter day backend - move to rest ee definitions to another package
-
-import './v1/engagementDashboard';
-import './v1/omnichannel';
-import './v1/sessions';
-import './v1/chat';
-import './v1/roles';
