--- apps/meteor/ee/definition/rest/v1/omnichannel/index.ts
@@ -1,3 +1,0 @@
-import './businessHours';
-import './businessUnits';
-import './cannedResponses';
