--- apps/meteor/ee/definition/rest/v1/sessions/SessionsProps.ts
@@ -1,18 +1,0 @@
-import Ajv from 'ajv';
-
-const ajv = new Ajv({ coerceTypes: true });
-
-export type SessionsProps = {
-	sessionId: string;
-};
-
-export const isSessionsProps = ajv.compile<SessionsProps>({
-	type: 'object',
-	properties: {
-		sessionId: {
-			type: 'string',
-		},
-	},
-	required: ['sessionId'],
-	additionalProperties: false,
-});
