+++ apps/meteor/ee/server/api/chat.ts
@@ -1,7 +1,23 @@
 import { Meteor } from 'meteor/meteor';
+import type { IMessage, ReadReceipt } from '@rocket.chat/core-typings';
 
 import { API } from '../../../app/api/server/api';
 import { hasLicense } from '../../app/license/server/license';
+
+type GetMessageReadReceiptsProps = {
+	messageId: IMessage['_id'];
+};
+
+declare module '@rocket.chat/rest-typings' {
+	// eslint-disable-next-line @typescript-eslint/naming-convention
+	interface Endpoints {
+		'/v1/chat.getMessageReadReceipts': {
+			GET: (params: GetMessageReadReceiptsProps) => {
+				receipts: ReadReceipt[];
+			};
+		};
+	}
+}
 
 API.v1.addRoute(
 	'chat.getMessageReadReceipts',
