+++ apps/meteor/ee/server/api/engagementDashboard/channels.ts
@@ -1,9 +1,36 @@
 import { check, Match } from 'meteor/check';
+import type { IDirectMessageRoom, IRoom } from '@rocket.chat/core-typings';
 
 import { API } from '../../../../app/api/server';
 import { findAllChannelsWithNumberOfMessages } from '../../lib/engagementDashboard/channels';
 import { isDateISOString, mapDateForAPI } from '../../lib/engagementDashboard/date';
 import { getPaginationItems } from '../../../../app/api/server/helpers/getPaginationItems';
+
+declare module '@rocket.chat/rest-typings' {
+	// eslint-disable-next-line @typescript-eslint/naming-convention
+	interface Endpoints {
+		'/v1/engagement-dashboard/channels/list': {
+			GET: (params: { start: string; end: string; offset?: number; count?: number }) => {
+				channels: {
+					room: {
+						_id: IRoom['_id'];
+						name: IRoom['name'] | IRoom['fname'];
+						ts: IRoom['ts'];
+						t: IRoom['t'];
+						_updatedAt: IRoom['_updatedAt'];
+						usernames?: IDirectMessageRoom['usernames'];
+					};
+					messages: number;
+					lastWeekMessages: number;
+					diffFromLastWeek: number;
+				}[];
+				count: number;
+				offset: number;
+				total: number;
+			};
+		};
+	}
+}
 
 API.v1.addRoute(
 	'engagement-dashboard/channels/list',
