+++ apps/meteor/ee/server/api/engagementDashboard/users.ts
@@ -1,4 +1,5 @@
 import { check, Match } from 'meteor/check';
+import type { IUser } from '@rocket.chat/core-typings';
 
 import { API } from '../../../../app/api/server';
 import {
@@ -9,6 +10,65 @@ import {
 	findUserSessionsByHourWithinAWeek,
 } from '../../lib/engagementDashboard/users';
 import { isDateISOString, transformDatesForAPI } from '../../lib/engagementDashboard/date';
+
+declare module '@rocket.chat/rest-typings' {
+	// eslint-disable-next-line @typescript-eslint/naming-convention
+	interface Endpoints {
+		'/v1/engagement-dashboard/users/active-users': {
+			GET: (params: { start: string; end: string }) => {
+				month: {
+					day: number;
+					month: number;
+					year: number;
+					usersList: IUser['_id'][];
+					users: number;
+				}[];
+			};
+		};
+		'/v1/engagement-dashboard/users/chat-busier/weekly-data': {
+			GET: (params: { start: string }) => {
+				month: {
+					users: number;
+					day: number;
+					month: number;
+					year: number;
+				}[];
+			};
+		};
+		'/v1/engagement-dashboard/users/chat-busier/hourly-data': {
+			GET: (params: { start: string }) => {
+				hours: {
+					users: number;
+					hour: number;
+				}[];
+			};
+		};
+		'/v1/engagement-dashboard/users/users-by-time-of-the-day-in-a-week': {
+			GET: (params: { start: string; end: string }) => {
+				week: {
+					users: number;
+					hour: number;
+					day: number;
+					month: number;
+					year: number;
+				}[];
+			};
+		};
+		'/v1/engagement-dashboard/users/new-users': {
+			GET: (params: { start: string; end: string }) => {
+				days: { day: Date; users: number }[];
+				period: {
+					count: number;
+					variation: number;
+				};
+				yesterday: {
+					count: number;
+					variation: number;
+				};
+			};
+		};
+	}
+}
 
 API.v1.addRoute(
 	'engagement-dashboard/users/new-users',
