+++ apps/meteor/ee/server/api/sessions.ts
@@ -1,18 +1,84 @@
 import { Users, Sessions } from '@rocket.chat/models';
-import type { IUser } from '@rocket.chat/core-typings';
+import type { IUser, ISession, DeviceManagementSession, DeviceManagementPopulatedSession } from '@rocket.chat/core-typings';
 import { escapeRegExp } from '@rocket.chat/string-helpers';
+import type { PaginatedResult, PaginatedRequest } from '@rocket.chat/rest-typings';
+import Ajv from 'ajv';
 
-import { isSessionsPaginateProps, isSessionsProps } from '../../definition/rest/v1/sessions';
 import { API } from '../../../app/api/server/api';
 import { hasLicense } from '../../app/license/server/license';
 import { Notifications } from '../../../app/notifications/server';
 import { getPaginationItems } from '../../../app/api/server/helpers/getPaginationItems';
+
+const ajv = new Ajv({ coerceTypes: true });
+
+type SessionsProps = {
+	sessionId: string;
+};
+
+const isSessionsProps = ajv.compile<SessionsProps>({
+	type: 'object',
+	properties: {
+		sessionId: {
+			type: 'string',
+		},
+	},
+	required: ['sessionId'],
+	additionalProperties: false,
+});
+
+type SessionsPaginateProps = PaginatedRequest<{
+	filter?: string;
+}>;
+
+const isSessionsPaginateProps = ajv.compile<SessionsPaginateProps>({
+	type: 'object',
+	properties: {
+		offset: {
+			type: 'number',
+		},
+		count: {
+			type: 'number',
+		},
+		filter: {
+			type: 'string',
+		},
+		sort: {
+			type: 'string',
+		},
+	},
+	required: [],
+	additionalProperties: false,
+});
 
 const validateSortKeys = (sortKeys: string[]): boolean => {
 	const validSortKeys = ['loginAt', 'device.name', 'device.os.name', 'device.os.version', '_user.name', '_user.username'];
 
 	return sortKeys.every((s) => validSortKeys.includes(s));
 };
+
+declare module '@rocket.chat/rest-typings' {
+	// eslint-disable-next-line @typescript-eslint/naming-convention
+	interface Endpoints {
+		'/v1/sessions/list': {
+			GET: (params: SessionsPaginateProps) => PaginatedResult<{ sessions: Array<DeviceManagementSession> }>;
+		};
+		'/v1/sessions/info': {
+			GET: (params: SessionsProps) => DeviceManagementSession;
+		};
+		'/v1/sessions/logout.me': {
+			POST: (params: SessionsProps) => Pick<ISession, 'sessionId'>;
+		};
+		'/v1/sessions/list.all': {
+			GET: (params: SessionsPaginateProps) => PaginatedResult<{ sessions: Array<DeviceManagementPopulatedSession> }>;
+		};
+		'/v1/sessions/info.admin': {
+			GET: (params: SessionsProps) => DeviceManagementPopulatedSession;
+		};
+		'/v1/sessions/logout': {
+			POST: (params: SessionsProps) => Pick<ISession, 'sessionId'>;
+		};
+	}
+}
 
 API.v1.addRoute(
 	'sessions/list',
